Initial implementation of t-digest for postgres
This change incoporates the rust t-digest implementation by Paul
Meng and Maarten Deprez (https://github.com/MnO2/t-digest) and builds
a postgres aggregate around it.  Note that the current version will
not compile as it requires some modification to the t-digest library
and will attempt to find the crate at a local path.