Fix 1b varlen handling
varlen types have a short version where there's only 1 byte of header
and everything is stored in-line. This breaks one of the assumptions
our value-wrapper is doing, so perform a copy in the case where we
detect our value is in that form to ensure that we have the full
4-byte header. I believe this is the correct trade off, as the
additional copy in the rare case that the value is small enough to use
this optimization /should/ be less expensive than the additional disk
space would be.