Add trigger disabling deps on experimental features
"experimental" features are just that, and there usage can be fraught:
we are neither capable, not desirous, of providing an upgrade path for
someone using experimental features to a newer version of this
extension, and we would prefer users to not become reliant on
experimental features. As such, our upgrade scripts for the extension are intended to contain a
'DROP SCHEMA timescale_analytics_experimental' at the start of the
script. To ensure users are adequately warned of this risk, this commit
adds a ddl_end trigger that prevents users from creating dependencies
on experimental features unless they set a guc indicating they know the
implications of what they're doing.