Fix bug in t-digest serialization
We were using the wrong attribute; the one we used specified that the
fields should be skipped during serialization, but would be found during
in the bytes during deserialization. What we want her is the attribute
that specifies the field should be skipped during serialization, and
initialized using 'Default::default()' during deserialization.