Remove ordered version, add tests, better naming
First pass addressing feedback and getting ready for an initial version to be
released on the experimental branch. We removed the ordered version of the
aggregate as it was redundant and not providing much value, made sure everything
fit in the experimental schema naming approach, added some initial testing to
make sure it works under normal conditions and renamed some functions based on
the feedback from reviewers.