Add LTTB implementation
Largest Triangle Three Buckets is a downsampling method that attempts to
preserve visual similarity with the original data. This commit adds a
simple implementation that ingests and sorts the entire data range. We
may eventually want to add a version that takes sorted input to better
handle large data sets.