Use newer Rust version in nightlies
We want to use a newer rust version to take advantage of cargo
improvements, and to use some of the newer panic handling functions.
The current version of Alpine docker is using has an older one in the
package manager, so we need to explicitly use the edge repo to get a
newer version. We will likely wish to continue using the edge repo so
we can update rust in the future without breaking nightlies.