Change base docker image to timescaledb-ha
This changes the base docker image to the fully featured docker image
`timescale/timescaledb-ha:pg13-latest`. This has a number of advantages
 1. It brings the nightly image more in-line with our eventual release
    channel.
 2. It brings the nightly image more in-line with the CI image, and
    improves the likelihood it'll be correct.
 3. The alpine-based Postgres image is a strange beast, with only
    partial support for collation (see issue #148 in the repo for more
    details). Switching to a Debian based image lets us avoid these
    issues.

It has the disadvantage that it results in a larger image, and we may
want to add the extension to the regular Alpine-based image in the
future to rectify this.