Add additional test runner
This PR also adds a test runner that can be run after post-install.rs to
check that tests still pass, and adds CI checks using this runner. This
should help us be confident that post-install.rs runs correctly.



Also re-ups the cache so we'll have cached versions of the new tools.