Update flat_serialize in tree and move in tree
This commit updates the flat-serialize version, and uses it's new trait
mechanism to enforce that all our flat-serialized types can in fact be
serialized like this. Right now there's still an unprincipled exception
for range.rs:DiskOption, we need to have one of optional fields or
generic types to remove it.

This commit also move flat-serialize it in-tree for ease-of maintenance;
now that flat_serialize is starting to infect the `crates` directory,
it's easier to just use a relative path to our own flat_serialize
everywhere.

Other quality-of-life improvements include by-value structs, thus
removing the need for lifetimes for many cases.