Display TSPoint timestamps as text
This commit adds textual serialization and deserialization to TSPoint
when it's being in/output to a human-readable format. Outputting the
integer representation of Postgres timestamps isn't particularly usable
to human beings, nor consumable by non-Postgres software, and this
work better for both vases.

The translation is done by an in-tree copy of timestamptz_in/out. Since
we only link to Postgres in the top-level extension crate, we currently
use undefined symbol hackery in crates/time-series to make this work.
Long-term, we should likely move it, and its transitive dependencies
into the extension crate.