Pipeline scaffolding
This commit adds the pipeline scaffolding; the basic code to run
pipeline elements. The commit also adds an lttb element to test it.

This commit also switches the timeseries text IO to to use a simple
explicit timeseries. I believe this is the most user-friendly default
for both human readers, and the simplest for downstream applications to
parse. We can always add a transform to send the more optimized layouts
if desired.