Add operator for user pipeline elements
This commit adds an even more experimental operator for user-defined
pipeline elements. This operator `|>>` can be used anywhere our other
operator can except that right side is a regproc instead of a pipeline
element eg. it looks like
```SQL
CREATE FUNCTION yesterday_squared(timeseries) RETURNS timeseries AS $$
    SELECT timeseries(time, value * value) \
    FROM (SELECT (unnest_series($1)).*) a \
    WHERE time>now() - '1 day'::interval;\
$$
;

SELECT series |>> 'yesterday_squared';
```

This commit creates a new operator instead of reusing the `|>` because
if we use `|>` for both this and and the regular timeseries elements
trying to do `series |> 'custom_element'` gets an ambiguous operator
error `timeseries |> unknown` is not unique. In the future we could
consider changing the element input function to fallback to checking if
the input is a regproc if it doesn't recognize it; the formats should be
different enough that there's no risk of collision.