Up pgx version
We're getting closer to being able to work on vanilla upstream.
Changes since last sync seem to alter the way the Spi client handles
timezones, so explicitly set the timezone on those tests that rely on
the old behavior.

Also update the readme and docker images to use the new branch to
install cargo-pgx even though this /technically/ isn't needed.