Update flat_serialize with owned versions
This commit updates flat_serialize with functions to turn a
flat_serialize type into an owned ('static) version of it. This is
useful when we need to return a guaranteed-owned version of a type.
It should generally be a nop to turn an already-owned instance into a
guaranteed-owned one, it's just a type change.