Remove InternalTimeSeries
Now that flat_serialize slices are CoW we can handle mutation "in-place"
and we don't need a separate type for it. This commit goes through and
removes the remaining references to the InternalTimeSeries type, and
goes a little overboard with en-passant refactoring.

The most notable refactor that isn't a direct result of removing the
InternalTimeSeries type is the rename and moving of the TimeSeries
iterator to a new `iter` submodule, along with some added
specializations to ensure that random-access to such iterators is as
efficient as we can make it.