Expunge MaybeOwnedTs
Now that TimeSeries handles ownership internally we no longer need to
track it in a container type. Besides making the code smaller and easier
to read, this also localizes all ownership info at the top level
function, hopefully making such code easier to write.
This code removes a number of now-redundant allocations, and gets us
closer to only reallocating the timeseries only when we actually need
to.