Add pipeline support function
This commit adds a pipeline support function that does right-associative
constant-folding of timeseries pipelines. Custom postgres operators are
always left-associative so pipeline such as
```
series |> foo() |> bar() |> baz() |> qux()
```
is equivalent to
```
((series |> foo()) |> bar()) |> baz() |> qux()
```
which will cause a new copy of the series to be materialized after every
pipeline element. This is much worse than the right-associative
```
series |> (foo() |> bar() |> baz() |> qux())
```
which would only require the series to be materialized twice: once on
input to the pipeline and once on pipeline completion.

This commit adds a support function that ensure pipelines always follow
the better execution path. It implements a peephole pass that looks for
expressions of the form
```
(series |> [constant-pipeline]) |> [other-constant-pipeline]
```
and transforms it into
```
series |> const([constant-pipeline] |> [other-constant-pipeline])
```
which should handle most cases.

(The one limitation is `map_data()` since this is currently marked as
`stable` it will not be turned into a const and thus prevents the
optimization. If this function does indeed need to be marked `stable`
instead of `immutable` this may be acceptable, but we could conceivably
write a function that handles this better.)