Changing serialization to use RON instead of JSON
This change modifies the text serialization form of toolkit objects
from JSON to RON (Rust Object Notation).  This is necessary due to
JSON not being able to express float point NaNs and infinities.  More
problematically, the presence of these values would result in us
writing out a string that we would be unable to serialize the object
back from.

This will break us trying to read in objects written out prior to this
change.  Luckily it is only the text based representation which is
affected, which is only used by Postgres dump and restore.  This does
mean, however, that dumps created prior to the version containing this
change will not be able to be restored from in the new version.  Users
are recommended to create new dumpfiles after upgrading.