Fix IO tests and add validation to accessors
It turns out we have a slight collision between the arrow accesors and
the function ones: when
```SQL
covariance('<some text>')
```
is called, Postgres naturally assumes that it is calling the function
```SQL
covariance(TEXT) -> Accessor
```
not
```SQL
covariance(StatsSummary2D, TEXT='sample') -> Double Precision
```
while this is more or less the behavior we want, it does break some IO
tests, and makes good error messaging for incorrect methods in the
accessors a higher priority. This commit fixes the IO tests and adds
validation to the accessors, unifying the method-detection code
en passant.