Make pipline-folding work with other finalizers
This commit changes how our pipeline-folding optimization works in order
to make it work with arbitrary finalizers. Previously, the optimization
worked by folding into the _first_ executor. That is, given a pipeline
such as
```SQL
series -1> foo() -2> bar() -3> baz()
```
the optimization would attempt to produce
```SQL
series -1> const(foo() -> bar() -> baz())
```
The first execution element `-1>` was special-cased to ensure that this
optimization was valid; the optimization only worked with the default
executor, but it was assumed others could be added later.

This commit changes the optimization to always fold into the latest
executor. That is given the same pipeline
```SQL
series -1> foo() -2> bar() -3> baz()
```
the optimization will now try to produce an execution like
```SQL
series -3> const(foo() -> bar() -> baz())
```
As long as we only apply the peephole optimization to `->`s that have a
type analogous to `(timeseries, Pipeline) RETURNS foo`, for some `foo`,
this transformation is valid, and removes a materialization of the
timeseries.

One downside to performing the optimization this way is that we lose the
fact that this execution is performed by an operator. This meas that in
cases where the expression is printed (for instance, an `EXPLAIN`) what
gets printed looks more like
```SQL
run_pipeline(series, const(foo() -> bar() -> baz()))
```
than the nice arrowed expression above. While this is less readable, I
believe the execution benefits outweigh this. Note that
the const-expanded
```SQL
const(foo() -> bar() -> baz())
```
is already fairly ugly.