Make datum caching slightly safer
We cache previously-flattened datums in our types so we don't need to
re-flatten every time if we pass an object to a succession of postgres
functions. Previously this was a little fraught as we had no way to
determine the provenance of the flattened datum: we could not tell if
it was a datum we allocated ourselves or one we received from postgres.
This commit adds this tracking and ensures we never return an inputted
datum unless we explicitly request to do so.