Walk dir in `post-install`'s `--dir`
This commit updates `post-intall --dir` to walk the directory looking
for the `.control` and `.so` files in order to find the correct
directories. It turns out to be difficult to manually determine which
directory to pass to `--dir`, and the format after the
`share`/`extension` might not be the same in every instance, so this
should fix both problems.