Reduce the likelihood of OOM aborts
By default rust will `abort()` the process when the allocator returns
NULL. Since many systems can't reliably determine when an allocation
will cause the process to run out of memory, and just rely on the OOM
killer cleaning up afterwards, this is acceptable for many workloads.
However, `abort()`-ing a Postgres will restart the database, and since
we often run Postgres on systems which _can_ reliably return NULL on
out-of-memory, we would like to take advantage of this to cleanly shut
down a single transaction when we fail to allocate. Long-term the
solution for this likely involves the `oom=panic` flag[1], but at the
time of writing the flag is not yet stable.

This allocator implements a partial solution for turning out-of-memory
into transaction-rollback instead of process-abort. It is a thin shim
over the System allocator that `panic!()`s when the System allocator
returns `NULL`. In the event that still have enough remaining memory to
serve the panic, this will unwind the stack all the way to
transaction-rollback. In the event we don't even have enough memory to
handle unwinding this will merely abort the process with a
panic-in-panic instead of a memory-allocation-failure. Under the
assumption that we're more likely to fail due to a few large allocations
rather than a very large number of small allocations, it seems likely
that we will have some memory remaining for unwinding, and that this
will reduce the likelihood of aborts.

[1] `oom=panic` issue: https://github.com/rust-lang/rust/issues/43596