Add Aggregate Builder
This commit adds an aggregate building macro for Toolkit, and migrates
TDigest and creates an Arbitrary aggregate to serve as demos.

The aggregate builder replaces our old way of constructing aggregates.
Instead of creating each function and manually writing the SQL to turn
them into an aggregate, the macro will do the SQL creation itself, and
also provide static typechecking to ensure the set of functions can be
used to create a valid aggregate. In particular it provides the
following benefits:
 1. It prevent the possibility that we get the inner transition state
    type wrong between different functions making up and aggregate. i.e.
    having a transition function `fn transition(Option<Inner<String>>)`
    with a final function `fn finally(Option<Inner<&str>>)`.
 2. It ensures that the aggregate context is switched to in,
    and only in, the correct places. (looking at array_agg, we want to
    be in the aggregate context of the transition and combine functions
    but _not_ the final function, node agg wants the result to be in the
    current expression context, so being in the aggregate context will
    just cause an additional allocation and copy.)
 3. Reduce the likelihood of errors in CREATE AGGREGATE statements.
    Though we will catch this in tests, it's annoying to get the error
    out.
 4. Focus aggregate reviewers on the important semantic components of
    the aggregate implementation instead of the boilerplate.
 5. Reduce the effort involved in writing aggregates.

The syntax we eventually settled on looks something like the following

```rust
// (we need a prefix so git doesn't treat the next line as a comment)
/**/ #[aggregate] impl aggregate_name
    type State = InternalTransitionType;

    fn transition(
        state: Option<Inner<State>>,
        // can have an arbitrary number of additional args using
        // #[sql_type("sql_type")] argument: RustType,
    ) -> Option<Inner<State>> {
        // transition function function body goes here
    }

    fn finally(state: Option<Inner<State>>) -> Option<ResultType> {
        // final function function body goes here
    }

    // the remaining items are optional

    // parallel-safety marker if desireable
    const PARALLEL_SAFE: bool = true;

    fn serialize(state: Inner<State>) -> bytea {
        // serialize function body goes here
    }

    fn deserialize(bytes: bytea) -> Inner<State> {
        // deserialize function body goes here
    }

    fn combine(state1: Option<Inner<State>>, state2: Option<Inner<State>>) -> Option<Inner<State>> {
        // combine function body goes here
    }
}
```

for example, the simplest aggregate, `arbitrary()`, that returns an
element from the aggregated values, could look like

```rust
    type State = String;

    fn transition(
        state: Option<Inner<State>>,
    ) -> Option<Inner<State>> {
        match state {
            Some(value) => Some(value),
            None => Some(value.into()),
        }
    }

    fn finally(mut state: Option<Inner<State>>) -> Option<String> {
        match &mut state {
            None => None,
            Some(state) => Some(state.clone()),
        }
    }
}
```

This syntax seems to be at a local optima: it looks reasonably familiar
to rust user by working analogously to an `impl block`; it is also
very similar to `CREATE AGGREGATE`'s syntax, so it should be easy to
determine what the SQL behavior will be; it has syntactic space for the
aggregate's name so it can be included naturally; and finally it should
be easily extensible to moving-aggregates, notably the moving aggregate
transition state should fit in as something like `type MovingState`.