Clean clippy lint warnings
This fixes any code in the crates or extension directory which generated warning when building with clippy.  A few cases which could not be cleaned have been explicitly allowed as follows:

Globally allowed:
modulo_one - This occurred in every pg_type when alignment checking the version field.
extra_unused_lifetimes - Clippy raised this in many places where the compiler required the lifetime.

Allowed in time_series.rs and hyperloglog.rs:
identity_op - Clippy raised this on flat_serialize enum types

Allowed individually on unsafe functions:
missing_safety_doc - These actually should be added in eventually, but is not part of this PR.