Adding serialization for FrequencyTransState
This change packages up the Datum serialization in datum_utils and
uses it to serialize the trans state for frequency aggregates.  It
also adds combine functionality to the FrequencyTransState, which is
used to allow the freq_agg to be considered parallel safe by postgres.

It also refactors the LocalAnyElement struct from frequency.rs
into it's own PgAnyElement under pg_any_element.rs.