Adjust typing to prevent errors
It's only the transition function that has anything like ownership
semantics, in the other functions the required behavior is more like a
borrow. This commit changes the aggregate builder to respect that, and
requires a borrow for the final, serialize, deserialize, and combine
functions.

We currently allow a mutable borrow for the serialize function
because t-digest requires it, we'll have to see if we keep allowing this in the
future as it worsens the type-error reporting a bit. There's also
questions as to what extent this is moral.