Elide `Inner<>` from surface signature
It turns out that we can in fact remove the `Inner<>`s from all of the
surface syntax with a little work. Most notably, we need to change the
true transition state from `Option<Inner<State>>` to
`Option<Inner<Option<State>>>` in order to handle failures in the
transition function correctly. The transition state needs to be freed
whenever the aggregate memory context is freed in order to prevent
leaks; we achieve this by having `Inner<State>`'s constructor free the
pointer `Inner` and it's contents `State` in a MemoryContext teardown
hook. However, if we have a signature like
```rust
fn transition(state: Option<State>) -> Option<State>
```
then during unwinding rust will free the `State` and trying to free it
again when the teardown hook fires will result in a double-free. In
order to prevent this, we instead store `Option<State>` to use as a
dynamic drop flag: in the wrapping function we do something like
```rust
fn transition_outer(inner: Option<Inner<Option<State>>>)
-> Option<Inner<Option<State>>> {
    let state: Option<State> = inner.as_mut()
        .map(|s| s.take())
        .flatten(); // 1
    let output = transition(state);
    match (inner, output) { // 2
        (None, None) => None,
        (None, state @ Some(..)) => Some(state.into),
        (Some(mut inner), state) => {
                *inner = state;
                Some(inner)
            }
    }
}
```
That is, we steal the pointee `State` leaving `None` behind at location
`// 1`. This means that if `transition(state)` unwinds, when we finally
tear down the aggregate memory context it'll only free the pointer, but
not recurse into `State`. After we finish the function, we check if
we've already allocated space to store it in, if we have we reuse it,
otherwise we allocate a new pointer to store the state in (location
`// 2`).