Remove the trigger
The trigger has a privilege escalation vulnerability due to the search
path not being locked down. Between this, the other headaches this has
given us around performance, and the fact that the core DB isn't using
a similar trigger for their experimental features, this is no longer
worth maintaining, so this commit deletes it.