Ensure old objects don't leak on update
This commit adds and update tests that ensures that after update we
there are no functions that still reference the old binary. We do this
by looking for functions that reference a binary other than the one for
the current version; since we version all our binaries like
`$libdir/`timescaledb_toolkit_x.y.z`, any old versions of the function
should reference a binary other than the current one.