Adding bigint and string freq specializations
This is a minimum change needed to specialize the frequency and topn aggregate
for integer and string types.  This change makes them much more user friendly,
but there are a lot of inefficiencies and duplicated code after this change.

The original AnyElement version of these aggregates can still be used as
raw_freq_agg or raw_topn_agg.