Upgrade to pgx 0.4.3
Went from 0.2.4 -> 0.3.0 -> 0.4.0 -> 0.4.3, editing the Cargo.toml,
building, and testing, and the experience was largely unremarkable.

Most of the changes come from followng the "Upgrading" section of the
[0.4.0](https://github.com/tcdi/pgx/releases/tag/v0.4.0) release notes.

Only other thing to note is that I ran into a *lot* of occurrences of
the following error:
```
 error[E0053]: method `input` has an incompatible type for trait
    --> extension/src/type_builder.rs:261:29
     |
 261 |             fn input(input: &std::ffi::CStr) -> $name<'input>
     |                             ^^^^^^^^^^^^^^^
     |                             |
     |                             expected struct `pgx::cstr_core::CStr`, found struct `std::ffi::CStr`
     |                             help: change the parameter type to match the trait: `&pgx::cstr_core::CStr`
```

Just blindly swapping in the `pgx::cstr_core` `CStr(ing)`s for the
`std::ffi` ones made Rust much happier.