Include both pgx 0.2.x and pgx 0.4.x in CI image
We're updating our pgx dependency to 0.4.3, and using `cargo-pgx
install` with pgx 0.4.x inside the update-tester causes problems.

Starting in 0.4, pgx changed it's SQL generation and gives the
following error when running `cargo pgx install` inside the update
tester with previous versions of the toolkit checked out:

  found `pgx` 0.2-0.3 series SQL generation while using `cargo-pgx`
  0.4 series.

As a solution, we're changing the docker image used in our CI setup to
include 0.2 series version of pgx installed at /pgx-0.2 as well as a
0.4 series version of pgx installed at /pgx-0.4.

With both pgx versions installed, for each toolkit release that is
checked out while running the update tester, we can infer the
appropriate cargo pgx subcommand to use by inspecting the Cargo.toml.