`update-tester` now requires new and old `cargo-pgx` arguments
The update tester now requires passing in the paths to "new" and "old"
installations of pgx in order to be able to invoke the correct version
of `cargo pgx`.

For each toolkit release that is checked out while running the update
tester, we can infer the appropriate cargo pgx subcommand to use by
inspecting the Cargo.toml.