remove PATH hack in tools/build
We added this in order for #408 to pass the CI tests. Our CI image had
pgx 0.2 on its PATH, but since we were upgrading to 0.4, we wanted to
use the 0.4 series pgx when running our CI tests (which use
`tools/build`).

This allowed us to pass our CI tests, but now that we've merged the
upgrade, we don't need this, and it surely isn't useful to anyone
running tests with `tools/build` locally.