Use layers in Dockerfile and add instructions
In order to speed up builds when developing the Dockerfile for CI
builds is layered. This will make it easier to work with when
developing since it does not require running all commands in one go and
can instead rely on the cached layers to avoid redoing work.

It also switches to using the official docker images from PGDG when
building rather than downloading them since this is what will be used
for the packages and for installations.

Also adds instructions on how to build a multi-platform build for the
CI as well as some troubleshooting tips.