Handle operators in update scripts
Now that we have stable operators we need to be able to update them.
Operators are slightly different than other objects in that we don't
schema-qualify them (they're too unpleasant to use when
schema-qualified). Instead we consider an operator unstable if one of
its inputs or its underlying function is unstable. Since operators are
create with the following syntax
```
CREATE OPERATOR <op> (
  PROCEDURE = <function name>,
  LEFTARG = <type name>,
  RIGHTARG = <type name>
);
```
we need to parse most of the CREATE to determine if the operator is
experimental.