Make tests pass in Canadian locales
Postgres date parsing is locale-dependent, so make dates parsed in tests
use a format that is parsed the same in (hopefully) all locales.

It would be better to set the locale of the database, but SETting the
LC_DATE parameter doesn't seem to have an impact. Ideally pgx should set
the database locale when creating a test database.