drop candlestick combine, serialize, and deserialize
We've determined that the cause for the bad results lives somewhere in
the functions that are used to support partial aggregation. We can at
least prevent folks from running the candlestick aggregates in
parallel mode and hitting this bug by dropping support for partial
aggregation until we've resolved the issue.