Push the post-release changes to main to a branch.
main is protected so we can't push directly to it.  We were considering
a pull request for this anyway, and have now decided to do it.

We'll figure out how to create the pull request later.  For now we can
do it manually.  At least we can stop making these edits manually.