Ensure PostgreSQL 16 compatibility
DateTimeParseError: PostgreSQL source code documents that we can pass
null pointers here:

 > extra can be NULL if not needed for the particular dterr value.
 > [...]
 > If escontext points to an ErrorSaveContext node, that is filled instead
 > of throwing an error

DecodeDateTime however does not allow us to pass a null pointer, reading
the source code, it is expected that extra is not null. Therefore, in
order for us to get this thing to compile, we create an empty
DateTimeErrorExtra, which gets passed along.