Ensure PostgreSQL 16 compatibility (#776)
* Ensure PostgreSQL 16 compatibility

DateTimeParseError: PostgreSQL source code documents that we can pass
null pointers here:

 > extra can be NULL if not needed for the particular dterr value.
 > [...]
 > If escontext points to an ErrorSaveContext node, that is filled instead
 > of throwing an error

DecodeDateTime however does not allow us to pass a null pointer, reading
the source code, it is expected that extra is not null. Therefore, in
order for us to get this thing to compile, we create an empty
DateTimeErrorExtra, which gets passed along.

* Ensure PostgreSQL 16 is tested

To achieve this, we have had to change quite a few things:

Explicitly list our Docker Image builds

The logic in the json was pretty complex, by using more yamls, and lists
we hope to achieve a workflow that is easier to inspect, reason with and
adapt.

We also need to reintroduce the difference between $PG_VERSIONS and
$TSDB_PG_VERSIONS, as PostgreSQL 16 support is not yet available for
TimescaleDB.

* Print debugging output on failure to start PG

* Explictly set materialized_only=false for test hypertable creations

* Skip some unsupported tests on PG16

* Skip PostgreSQL 16 doc tests

---------

Co-authored-by: syvb <smitty@timescale.com>