fix: correct broken serialization
Unfortunately the usage of `Internal` in function APIs throws type
safety out of the window. A handful of aggregate serialization
functions were incorrectly implemented, and operated on the incorrect
type, resulting in incorrect data in the serialized type.

Fortunately, this serialization is only used when [partial aggregation]
kicks in, so is not always triggered.

[partial aggregation]: https://www.postgresql.org/docs/current/xaggr.html#XAGGR-PARTIAL-AGGREGATES