Use a Vec and sorting to compact the SketchHashMap
Profiling showed us that we spend a lot of time in the compact function
(60%+ for the current test suite). We therefore want to reduce:

- malloc calls
- cpu usage in general

We do so by introducing a Vec that is only used during the compaction
itself, the actual HashMap will be reused every time. The Vec itself
will also stick around once it has been allocated.

We also instead of walking through the HashMap by following the linked
list, drain the HashMap, which should reduce the amount of comparisons
required, and therefore reduce CPU