Replace entry() with entry_upsert()
Profiling showed that this function is quite the hotspot. By changing
the implementation slightly, instead of walking the tree using the
Linked List, but iterate directly over the values, we improve the
throughput of certain CPU bound queries.

We've seen a reduction in time needed of > 50% for certain rollup
queries.

Due to the way entry() was called, and the way the Borrow Checker is
unable to help us keep 2 mutable references into a map, we were doing
double lookups into the backing HashMap pretty much always when this
function was called.

However, looking at the code, the only callers of this function only
wanted to either increment by 1, or by a count.

Therefore, make a function to actually support that usecase, which
doesn't have this problem with the Borrow Checker, as it doesn't have to
return a mutable reference: It actually does the work immediately